+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `emailAttachmentsWebhook` to email form workflow events validation
+
 ## [5.1.3] - 2023-12-12
 
 ### Fixed