+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Updated
 
+- Form validation to include `customCssClasses` prop
 - `forms.getSubmissionData` to return undefined when an `AccessDenied` error is thrown
 
 ### Added