+++ CHANGELOG.md
@@ -10,9 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Updated
 
 - PDF config validation to include PDF Page size
-
-### Updated
-
+- Form validation to include `customCssClasses` prop
 - `forms.getSubmissionData` to return undefined when an `AccessDenied` error is thrown
 
 ### Added